--- apps/laboratory/src/components/NetworksButton.tsx
@@ -1,3 +1,0 @@
-export function NetworksButton() {
-  return <w3m-network-button />
-}
