+++ apps/laboratory/src/components/Web3ModalButtons.tsx
@@ -1,0 +1,8 @@
+export function Web3ModalButtons() {
+  return (
+    <>
+      <w3m-button />
+      <w3m-network-button />
+    </>
+  )
+}
