--- apps/laboratory/src/constants.ts
@@ -1,3 +1,0 @@
-export const SigningSucceededToastTitle = 'Signing Succeeded'
-export const SigningFailedToastTitle = 'Signing Failed'
-export const TestIdSiweAuthenticationStatus = 'w3m-authentication-status'
