+++ apps/laboratory/src/utils/ConstantsUtil.ts
@@ -1,0 +1,5 @@
+export const ConstantsUtil = {
+  SigningSucceededToastTitle: 'Signing Succeeded',
+  SigningFailedToastTitle: 'Signing Failed',
+  TestIdSiweAuthenticationStatus: 'w3m-authentication-status'
+}
