+++ apps/laboratory/src/utils/WagmiConstants.ts
@@ -1,0 +1,31 @@
+import {
+  arbitrum,
+  aurora,
+  avalanche,
+  base,
+  bsc,
+  celo,
+  gnosis,
+  mainnet,
+  optimism,
+  polygon,
+  zkSync,
+  zora
+} from 'wagmi/chains'
+
+export const WagmiConstantsUtil = {
+  chains: [
+    mainnet,
+    arbitrum,
+    polygon,
+    avalanche,
+    bsc,
+    optimism,
+    gnosis,
+    zkSync,
+    zora,
+    base,
+    celo,
+    aurora
+  ]
+}
